requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
openai>=1.0.0
pandas>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
streamlit>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0